feat: add feature-structure rule to specs config and move feature-root files
