+++ apps/eclair/src/features/domain-map/queries/edge-aggregation.ts
