+++ apps/eclair/src/features/domain-map/queries/external-domain-handling.ts
