+++ apps/eclair/src/features/domain-map/queries/extract-domain-map.reactflow.spec.ts
@@ -1,7 +1,7 @@
 import {
   describe, it, expect 
 } from 'vitest'
-import { extractDomainMap } from '../extract-domain-map'
+import { extractDomainMap } from './extract-domain-map'
 import type { RiviereGraph } from '@living-architecture/riviere-schema'
 import {
   parseNode,
