+++ apps/eclair/src/features/domain-map/queries/extract-domain-map.ts
@@ -1,7 +1,307 @@
-export {
-  extractDomainMap,
-  getConnectedDomains,
-  type ConnectionDetail,
-  type DomainNodeData,
-  type DomainEdgeData,
-} from '../extract-domain-map'
+import type { RiviereGraph } from '@living-architecture/riviere-schema'
+import type {
+  Node, Edge 
+} from '@xyflow/react'
+import dagre from 'dagre'
+import { getClosestHandle } from '@/platform/infra/layout/handle-positioning'
+import { RiviereQuery } from '@living-architecture/riviere-query'
+import {
+  formatEdgeLabel, type ConnectionDetail, type EdgeAggregation 
+} from './edgeAggregation'
+import { LayoutError } from '@/platform/infra/errors/errors'
+import { aggregateDomainEdges } from './edge-aggregation'
+import {
+  aggregateExternalEdges, createExternalNodeId 
+} from './external-domain-handling'
+
+export type { ConnectionDetail } from './edgeAggregation'
+
+const LABEL_BG_PADDING: [number, number] = [4, 6]
+const DOMAIN_NODE_SIZE = 120
+const EXTERNAL_NODE_SIZE = 100
+
+export type { DomainNodeData } from '@/platform/domain/domain-node-types'
+
+export interface DomainEdgeData {
+  apiCount: number
+  eventCount: number
+  connections: ConnectionDetail[]
+}
+
+export type DomainNode = Node<DomainNodeData, 'domain'>
+export type DomainEdge = Edge<DomainEdgeData>
+
+interface DomainMapData {
+  domainNodes: DomainNode[]
+  domainEdges: DomainEdge[]
+}
+
+interface LayoutInput {
+  domainIds: string[]
+  edges: Array<{
+    source: string
+    target: string
+  }>
+  nodeSizes: Map<string, number>
+}
+
+function computeDagreLayout(input: LayoutInput): Map<
+  string,
+  {
+    x: number
+    y: number
+  }
+> {
+  const g = new dagre.graphlib.Graph().setDefaultEdgeLabel(() => ({}))
+
+  g.setGraph({
+    rankdir: 'LR',
+    nodesep: 40,
+    ranksep: 80,
+    marginx: 20,
+    marginy: 20,
+  })
+
+  for (const domainId of input.domainIds) {
+    const size = input.nodeSizes.get(domainId)
+    if (size === undefined) {
+      throw new LayoutError(`Domain ${domainId} missing from nodeSizes`)
+    }
+    g.setNode(domainId, {
+      width: size,
+      height: size,
+    })
+  }
+
+  for (const edge of input.edges) {
+    g.setEdge(edge.source, edge.target)
+  }
+
+  dagre.layout(g)
+
+  const positions = new Map<
+    string,
+    {
+      x: number
+      y: number
+    }
+  >()
+  for (const domainId of input.domainIds) {
+    const node = g.node(domainId)
+    positions.set(domainId, {
+      x: node.x,
+      y: node.y,
+    })
+  }
+
+  return positions
+}
+
+export function extractDomainMap(graph: RiviereGraph): DomainMapData {
+  const query = new RiviereQuery(graph)
+  const externalDomains = query.externalDomains()
+
+  const domainCounts = new Map<string, number>()
+  for (const node of graph.components) {
+    const currentCount = domainCounts.get(node.domain)
+    const count = currentCount ?? 0
+    domainCounts.set(node.domain, count + 1)
+  }
+
+  const nodeInfo = new Map<
+    string,
+    {
+      domain: string
+      name: string
+      type: string
+    }
+  >()
+  for (const node of graph.components) {
+    nodeInfo.set(node.id, {
+      domain: node.domain,
+      name: node.name,
+      type: node.type,
+    })
+  }
+
+  const edgeAggregation = new Map<string, EdgeAggregation>()
+  aggregateDomainEdges(graph.links, nodeInfo, edgeAggregation)
+
+  const externalEdges = aggregateExternalEdges(graph, nodeInfo)
+
+  const domains = Array.from(domainCounts.entries())
+  const externalNodeIds = externalDomains.map((ed) => createExternalNodeId(ed.name))
+  const allNodeIds = [...domains.map(([domain]) => domain), ...externalNodeIds]
+
+  const layoutEdges = [
+    ...Array.from(edgeAggregation.values()).map((agg) => ({
+      source: agg.source,
+      target: agg.target,
+    })),
+    ...externalEdges.map((e) => ({
+      source: e.sourceDomain,
+      target: createExternalNodeId(e.targetName),
+    })),
+  ]
+
+  const nodeSizes = new Map<string, number>()
+  for (const [domain] of domains) {
+    nodeSizes.set(domain, DOMAIN_NODE_SIZE)
+  }
+  for (const ed of externalDomains) {
+    nodeSizes.set(createExternalNodeId(ed.name), EXTERNAL_NODE_SIZE)
+  }
+
+  const domainPositions = computeDagreLayout({
+    domainIds: allNodeIds,
+    edges: layoutEdges,
+    nodeSizes,
+  })
+
+  const domainNodes: DomainNode[] = domains.map(([domain, nodeCount]) => {
+    const position = domainPositions.get(domain)
+    if (position === undefined) {
+      throw new LayoutError(`Domain ${domain} missing from layout computation`)
+    }
+    const calculatedSize = nodeSizes.get(domain)
+    return {
+      id: domain,
+      type: 'domain',
+      position,
+      data: {
+        label: domain,
+        nodeCount,
+        calculatedSize,
+        isExternal: false,
+      },
+    }
+  })
+
+  const externalNodes: DomainNode[] = externalDomains.map((ed) => {
+    const nodeId = createExternalNodeId(ed.name)
+    const position = domainPositions.get(nodeId)
+    if (position === undefined) {
+      throw new LayoutError(`External domain ${ed.name} missing from layout computation`)
+    }
+    const calculatedSize = nodeSizes.get(nodeId)
+    return {
+      id: nodeId,
+      type: 'domain',
+      position,
+      data: {
+        label: ed.name,
+        nodeCount: ed.connectionCount,
+        calculatedSize,
+        isExternal: true,
+      },
+    }
+  })
+
+  const allNodes = [...domainNodes, ...externalNodes]
+
+  const domainEdges: DomainEdge[] = Array.from(edgeAggregation.entries()).map(([key, agg]) => {
+    const sourcePos = domainPositions.get(agg.source)
+    const targetPos = domainPositions.get(agg.target)
+    if (sourcePos === undefined || targetPos === undefined) {
+      throw new LayoutError(
+        `Edge references missing domain position: source=${agg.source} target=${agg.target}`,
+      )
+    }
+    const handles = getClosestHandle(sourcePos, targetPos)
+
+    const isEventOnly = agg.eventCount > 0 && agg.apiCount === 0
+    const strokeColor = isEventOnly ? '#F59E0B' : '#06B6D4'
+    const arrowMarker = isEventOnly ? 'url(#arrow-amber)' : 'url(#arrow-cyan)'
+
+    return {
+      id: key,
+      source: agg.source,
+      target: agg.target,
+      sourceHandle: handles.sourceHandle,
+      targetHandle: handles.targetHandle,
+      label: formatEdgeLabel(agg.apiCount, agg.eventCount),
+      data: {
+        apiCount: agg.apiCount,
+        eventCount: agg.eventCount,
+        connections: agg.connections,
+      },
+      style: { stroke: strokeColor },
+      labelStyle: {
+        fontSize: 10,
+        fontWeight: 600,
+        fill: '#1f2937',
+      },
+      labelBgStyle: {
+        fill: 'rgba(255, 255, 255, 0.85)',
+        stroke: 'rgba(0, 0, 0, 0.1)',
+        strokeWidth: 1,
+        filter: 'drop-shadow(0 1px 2px rgba(0, 0, 0, 0.1))',
+      },
+      labelBgPadding: LABEL_BG_PADDING,
+      labelBgBorderRadius: 4,
+      markerEnd: arrowMarker,
+    }
+  })
+
+  const externalEdgesForMap: DomainEdge[] = externalEdges.map((e) => {
+    const targetId = createExternalNodeId(e.targetName)
+    const sourcePos = domainPositions.get(e.sourceDomain)
+    const targetPos = domainPositions.get(targetId)
+    if (sourcePos === undefined || targetPos === undefined) {
+      throw new LayoutError(
+        `External edge missing position: source=${e.sourceDomain} target=${targetId}`,
+      )
+    }
+    const handles = getClosestHandle(sourcePos, targetPos)
+
+    return {
+      id: `${e.sourceDomain}->${targetId}`,
+      source: e.sourceDomain,
+      target: targetId,
+      sourceHandle: handles.sourceHandle,
+      targetHandle: handles.targetHandle,
+      label: `${e.connectionCount} API`,
+      data: {
+        apiCount: e.connectionCount,
+        eventCount: 0,
+        connections: e.connections,
+      },
+      style: {
+        stroke: '#F97316',
+        strokeDasharray: '5,5',
+      },
+      labelStyle: {
+        fontSize: 10,
+        fontWeight: 600,
+        fill: '#1f2937',
+      },
+      labelBgStyle: {
+        fill: 'rgba(255, 255, 255, 0.85)',
+        stroke: 'rgba(0, 0, 0, 0.1)',
+        strokeWidth: 1,
+        filter: 'drop-shadow(0 1px 2px rgba(0, 0, 0, 0.1))',
+      },
+      labelBgPadding: LABEL_BG_PADDING,
+      labelBgBorderRadius: 4,
+      markerEnd: 'url(#arrow-orange)',
+    }
+  })
+
+  return {
+    domainNodes: allNodes,
+    domainEdges: [...domainEdges, ...externalEdgesForMap],
+  }
+}
+
+export function getConnectedDomains(domain: string, edges: DomainEdge[]): Set<string> {
+  const connected = new Set<string>()
+  for (const edge of edges) {
+    if (edge.source === domain) {
+      connected.add(edge.target)
+    }
+    if (edge.target === domain) {
+      connected.add(edge.source)
+    }
+  }
+  return connected
+}
