+++ apps/eclair/src/features/flows/entrypoint/FlowsPage.spec.tsx
@@ -6,7 +6,7 @@ import {
 } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 import userEvent from '@testing-library/user-event'
-import { FlowsPage } from './entrypoint/FlowsPage'
+import { FlowsPage } from './FlowsPage'
 import type { RiviereGraph } from '@living-architecture/riviere-schema'
 import {
   parseNode, parseEdge, parseDomainMetadata
