+++ knip.json
@@ -46,7 +46,7 @@
         "features/get-pr-feedback/entrypoint/cli.ts",
         "features/respond-to-feedback/entrypoint/cli.ts",
         "features/claude-hooks/entrypoint/hook-router.ts",
-        "features/claude-hooks/claude-hooks.spec.ts",
+        "features/claude-hooks/domain/claude-hooks.spec.ts",
         "shell/index.ts"
       ],
       "project": [
